Bugfix #6607 - Records with large long values do not sort correctly
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk@698 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/data/LongDataField.java
@@ -405,7 +405,7 @@ public class LongDataField extends DataField implements Numeric, Comparable{
 		if (isNull) return -1;
 		
 		if (obj instanceof LongDataField){
-			return compareTo(((LongDataField) obj).getInt());
+			return compareTo(((LongDataField) obj).getLong());
 		}else if (obj instanceof Integer){
 			return compareTo(((Integer)obj).longValue());
 		}else if (obj instanceof Long){
